{"schema":"decodex/commit/1","summary":"Modularize frozen toolbar rendering","authority":"manual"}
